requirements.txt
appdirs==1.4.4
appnope==0.1.3
asttokens==2.0.8
audioread==3.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
blis==0.7.8
boto3==1.24.62
botocore==1.27.62
bs4==0.0.1
catalogue==2.0.8
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cssselect==1.1.0
cycler==0.11.0
cymem==2.0.6
decorator==4.4.2
distlib==0.3.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
executing==1.0.0
fake-useragent==0.1.11
ffmpeg==1.4
filelock==3.8.0
Flask==2.2.2
fonttools==4.37.1
huggingface-hub==0.9.1
idna==3.3
imageio==2.21.2
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
kiwisolver==1.4.4
langcodes==3.3.0
librosa==0.9.2
llvmlite==0.39.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
moviepy==1.0.3
murmurhash==1.0.8
nltk==3.7
numba==0.56.0
numpy==1.22.0
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pafy==0.5.5
parse==1.19.0
parso==0.8.3
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pooch==1.6.0
preshed==3.0.7
proglog==0.1.10
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.9.2
pyee==8.2.2
Pygments==2.13.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
python-vlc==3.0.16120
pytube==12.1.0
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
requests-html==0.10.0
requests-oauthlib==1.3.1
resampy==0.4.0
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
smart-open==5.2.1
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SpeechRecognition==3.8.1
srsly==2.4.4
stack-data==0.5.0
thinc==8.1.0
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.0
traitlets==5.3.0
transformers==4.21.2
tweepy==4.10.1
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.4
w3lib==2.0.1
wasabi==0.10.1
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.2.2
wxPython==4.2.0
youtube-dl==2021.12.17
zipp==3.8.1
